requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
config==0.5.1
executor==23.2
fasteners==0.19
frozenlist==1.4.1
greenlet==3.0.3
humanfriendly==10.0
idna==3.6
loguru==0.7.2
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
property-manager==3.0
psycopg2-binary==2.9.9
pydantic==1.10.14
pyreadline3==3.4.1
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
verboselogs==1.7
win32-setctime==1.1.0
yarl==1.9.4
